refactor: rename markdown.ts to report.ts
